Extract SSR setup locale middleware
