+++ client/lib/get-asset-file-path/index.js
@@ -1,0 +1,8 @@
+/**
+ * External dependencies
+ */
+import path from 'path';
+
+export default function getAssetFilePath( target, filePath ) {
+	return path.join( __dirname, '..', '..', '..', 'public', target, filePath );
+}
