+++ client/my-sites/theme/index.node.js
@@ -1,9 +1,8 @@
 /**
  * Internal dependencies
  */
-import { makeLayout } from 'calypso/controller';
+import { makeLayout, ssrSetupLocale } from 'calypso/controller';
 import { details, fetchThemeDetailsData, notFoundError } from './controller';
-import { setupLocale } from 'calypso/my-sites/themes';
 import { getLanguageRouteParam } from 'calypso/lib/i18n-utils';
 
 export default function ( router ) {
@@ -12,7 +11,7 @@ export default function ( router ) {
 	router( '/theme', ( { res } ) => res.redirect( '/themes' ) );
 	router(
 		`/${ langParam }/theme/:slug/:section(setup|support)?/:site_id?`,
-		setupLocale,
+		ssrSetupLocale,
 		fetchThemeDetailsData,
 		details,
 		makeLayout,
