+++ client/my-sites/themes/index.node.js
@@ -1,16 +1,7 @@
 /**
- * External dependencies
- */
-/* eslint-disable import/no-nodejs-modules */
-import fs from 'fs';
-import path from 'path';
-/* eslint-enable import/no-nodejs-modules */
-import { localesToSubdomains } from '@automattic/i18n-utils';
-
-/**
  * Internal dependencies
  */
-import { makeLayout } from 'calypso/controller';
+import { makeLayout, ssrSetupLocale } from 'calypso/controller';
 import {
 	fetchThemeData,
 	fetchThemeFilters,
@@ -20,60 +11,7 @@ import {
 	redirectToThemeDetails,
 } from './controller';
 import { validateFilters, validateVertical } from './validate-filters';
-import { getLanguage, getLanguageRouteParam } from 'calypso/lib/i18n-utils';
-import { setLocaleRawData } from 'calypso/state/ui/language/actions';
-
-const translationsCache = {};
-
-export function setupLocale( context, next ) {
-	if ( ! context.params.lang ) {
-		const localeDataPlaceholder = { '': {} };
-		context.store.dispatch( setLocaleRawData( localeDataPlaceholder ) ); // Reset to default locale
-
-		next();
-		return;
-	}
-	const subdomainsToLocales = Object.entries( localesToSubdomains ).reduce(
-		( acc, [ key, value ] ) => {
-			return {
-				...acc,
-				[ value ]: key,
-			};
-		},
-		{}
-	);
-	const langSlug = subdomainsToLocales[ context.params.lang ] || context.params.lang;
-	const language = getLanguage( langSlug );
-
-	if ( language ) {
-		context.lang = language.langSlug;
-		context.isRTL = language.rtl ? true : false;
-
-		if ( typeof translationsCache[ context.lang ] === 'undefined' ) {
-			translationsCache[ context.lang ] = JSON.parse(
-				fs.readFileSync(
-					path.join(
-						__dirname,
-						'..',
-						'..',
-						'..',
-						'public',
-						'evergreen',
-						'languages',
-						`${ context.lang }-v1.1.json`
-					),
-					'utf-8'
-				)
-			);
-		}
-
-		const translations = translationsCache[ context.lang ];
-
-		context.store.dispatch( setLocaleRawData( translations ) );
-	}
-
-	next();
-}
+import { getLanguageRouteParam } from 'calypso/lib/i18n-utils';
 
 export default function ( router ) {
 	// Redirect interim showcase route to permanent one
@@ -91,7 +29,7 @@ export default function ( router ) {
 	];
 	router(
 		showcaseRoutes,
-		setupLocale,
+		ssrSetupLocale,
 		fetchThemeFilters,
 		validateVertical,
 		validateFilters,
